nstats: introduce templating to improve style-ability
So far, nsstats.tcl was a singe-file package, which is convenient for
quick installations. However, the UI made a dusty and outdated
impression, not providing any kind of stylability.  The new version
comes with a template "nsstats.adp" which works for plain NaviServer
instances and OpenACS.

In order to keep the simplicity of earlier versions, a default
template is provided in the nsstats.tcl file, which is used, when no
file "nsstats.adp" is found. So, the single-file copy is still
possible, but not recommended.

The provided "nsstats.adp" file uses the CSS classes from
https://www.w3schools.com/w3css/

Furthermore, the new version provides a graphical UI for "ns_server
-server SERVER requestprocs", which is linked from the "Process" page
for every server.

bumped version number to 1.12